fix(PaymentMethodRequest.java): fix totalAmount data type to Long
✨ Changed totalAmount data type from Integer to Long for compatibility.

Main

+++ src/main/java/it/gov/pagopa/afm/calculator/model/paymentmethods/PaymentMethodRequest.java
@@ -29,7 +29,7 @@ public class PaymentMethodRequest {
     String bin;
 
     @NotNull
-    Integer totalAmount;
+    Long totalAmount;
 
     @NotNull
     @NotEmpty
